backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openai==1.107.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
SQLAlchemy==2.0.48
Werkzeug==3.1.7
